feat(cloud): group generated notebook runs
